split out tests in line with validation split
